Moved mock data to agreed upon data structure folder, Added PlatformFilter, updated FilterContext to include Platform Filter and Mock Data
